+++ kpal-plugin/src/errors.rs
@@ -1,6 +1,6 @@
 //! Structures that provide error information to clients of the plugin library.
 
-pub static ERRORS: [&[u8]; 8] = [
+pub static ERRORS: [&[u8]; 9] = [
     // 0 PLUGIN_OK
     b"Plugin OK\0",
     // 1 UNDEFINED_ERR
@@ -11,11 +11,13 @@ pub static ERRORS: [&[u8]; 8] = [
     b"Attribute does not exist\0",
     // 4 ATTRIBUTE_TYPE_MISMATCH
     b"Attribute types do not match\0",
-    // 5 IO_ERR
+    // 5 ATTRIBUTE_IS_NOT_SETTABLE
+    b"Attribute cannot be set\0",
+    // 6 IO_ERR
     b"IO operation failed\0",
-    // 6 NUMERIC_CONVERSION_ERR
+    // 7 NUMERIC_CONVERSION_ERR
     b"Could not convert numeric value into a different type\0",
-    // 7 NULL_PTR_ERR
+    // 8 NULL_PTR_ERR
     b"The plugin encountered a null pointer\0",
 ];
 
@@ -28,7 +30,8 @@ pub mod constants {
     pub const PLUGIN_INIT_ERR: c_int = 2;
     pub const ATTRIBUTE_DOES_NOT_EXIST: c_int = 3;
     pub const ATTRIBUTE_TYPE_MISMATCH: c_int = 4;
-    pub const IO_ERR: c_int = 5;
-    pub const NUMERIC_CONVERSION_ERR: c_int = 6;
-    pub const NULL_PTR_ERR: c_int = 7;
+    pub const ATTRIBUTE_IS_NOT_SETTABLE: c_int = 5;
+    pub const IO_ERR: c_int = 6;
+    pub const NUMERIC_CONVERSION_ERR: c_int = 7;
+    pub const NULL_PTR_ERR: c_int = 8;
 }